+++ docs/release_notes.rst
@@ -5,6 +5,9 @@
 Version 0.3.2 (under development)
 ---------------------------------
 
+* Fixed ``verify`` method of IPN/PDT so that it can be re-run in the case
+  of a PayPal server error.
+
 Version 0.3.1
 -------------